feat: remove extra _state constant from project
